refactor(Components): Moving components
Moving the single source of truth to inside app directory
