+++ __tests__/index.ios.js
@@ -1,7 +1,7 @@
 import 'react-native'
 import React from 'react'
 import renderer from 'react-test-renderer'
-import Index from '../index.ios'
+import Index from '../app'
 
 it('renders correctly', () => {
   renderer.create(
